SofT-GRPO-master-main/verl-0.4.x/requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft>=0.7.1  # Required for transformers 4.46.0+ compatibility
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict<=0.6.2
torchdata
transformers>=4.46.0  # Required for peft compatibility (modeling_layers module)
# vllm==0.8.4
wandb
packaging>=20.0
uvicorn
fastapi
